fix add post

+++ Back-end/controllers/posts.js
@@ -26,9 +26,9 @@ exports.createPost = (req, res) => {
       }`,
     });
     post
-    .save()
-    .then(() => res.status(201).json(post))
-    .catch((error) => res.status(400).json({ error }));
+      .save()
+      .then(() => res.status(201).json(post))
+      .catch((error) => res.status(400).json({ error }));
   } else {
     console.log(req.file);
     const post = new Post({
@@ -61,12 +61,18 @@ exports.getPostsById = (req, res) => {
 exports.deletePost = (req, res) => {
   Post.findOne({ _id: req.params.id })
     .then((post) => {
-      const filename = post.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
+      if (post.imageUrl) {
+        const filename = post.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Post.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
         Post.deleteOne({ _id: req.params.id })
           .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
           .catch((error) => res.status(400).json({ error }));
-      });
+      }
     })
     .catch((error) => res.status(500).json({ error }));
 };
